dashboard-crea-mcp/mcp-server/requirements.txt
mcp>=0.1.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
litellm>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sqlalchemy>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
